requirements.txt
pyperclip
pyobjc-core
pyobjc-framework-CoreServices